+++ src/types/type.ts
@@ -1,0 +1,6 @@
+export interface analyzeDataType {
+	category: string
+	detail: string
+	price: number
+	date: string
+}
